% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_buoy}
\alias{get_buoy}
\title{Retrieve buoy data for one or more buoys}
\usage{
get_buoy(x = dplyr::slice(read_buoy_list(), 1:2), ...)
}
\arguments{
\item{x}{a table of buoy metadata for one or more buoys}

\item{...}{other arguments for the [buoydata::get_buoy_data()]}
}
\value{
table
}
\description{
Retrieve buoy data for one or more buoys
}
