% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_buoy_list}
\alias{read_buoy_list}
\title{Read a list of buoy info, and give user the option to make it a spatial
object.}
\usage{
read_buoy_list(form = c("table", "sf")[1])
}
\arguments{
\item{form}{one of "table" or spatial table ("sf")}
}
\value{
table or an sf table
}
\description{
Read a list of buoy info, and give user the option to make it a spatial
object.
}
