% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_buoy}
\alias{write_buoy}
\title{Write buoy data to file}
\usage{
write_buoy(x = read_buoy(), filename = "buoy.csv.gz")
}
\arguments{
\item{x}{table of data}

\item{filename}{str/chr, the name of the file to write}
}
\value{
the input table
}
\description{
Write buoy data to file
}
