% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bauja-package.R
\docType{package}
\name{bauja}
\alias{bauja-package}
\alias{bauja}
\title{bauja package}
\description{
A R package for accessing and manipulating buoy data from NDBC.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/btupper/bauja}
}

}
\author{
\strong{Maintainer}: Tupper Ben \email{btupper@bigelow.org}

}
